+++ src/BC-TGV.f90
@@ -47,10 +47,6 @@ module tgv
   character(len=100) :: fileformat
   character(len=1),parameter :: NL=char(10) !new line character
 
-  !probes !só vai funcionar se a impressão for em relação ao lapis X!
-  integer :: nprobes
-  integer, save, allocatable, dimension(:) :: rankprobes, nxprobes, nyprobes, nzprobes
-
   PRIVATE ! All functions/subroutines private by default
   PUBLIC :: init_tgv, boundary_conditions_tgv, postprocess_tgv